+++ src/main/groovy/org/folio/inventory/domain/ingest/IngestMessages.groovy
@@ -11,6 +11,9 @@ class IngestMessages {
     Map materialTypes,
     Map loanTypes,
     Map locations,
+    Map identifierTypes,
+    Map instanceTypes,
+    Map creatorTypes,
     jobId,
     Context context) {
 
@@ -20,7 +23,10 @@ class IngestMessages {
       .put("records", records)
       .put("materialTypes", materialTypes)
       .put("loanTypes", loanTypes)
-      .put("locations", locations))
+      .put("locations", locations)
+      .put("identifierTypes", identifierTypes)
+      .put("instanceTypes", instanceTypes)
+      .put("creatorTypes", creatorTypes))
   }
 
   static completed(jobId, Context context) {
@@ -35,5 +41,4 @@ class IngestMessages {
      "token"     : context.token,
      "okapiLocation": context.okapiLocation]
   }
-
 }