+++ src/test/groovy/org/folio/inventory/parsing/ModsParsingExamples.groovy
@@ -32,6 +32,8 @@ class ModsParsingExamples {
     assert california.identifiers.size() == 1
     assert hasIdentifier(california, "UkMaC", "69228882")
 
+    assert california.creators.size() == 1
+    hasCreator(california, "Huntley, Henry Veel")
 
     def studien = records.find({
       similarTo(it, "Studien zur Geschichte der Notenschrift.", "69247446")
@@ -41,6 +43,9 @@ class ModsParsingExamples {
 
     assert studien.identifiers.size() == 1
     assert hasIdentifier(studien, "UkMaC", "69247446")
+
+    assert studien.creators.size() == 1
+    hasCreator(studien, "Riemann, Karl Wilhelm J. Hugo.")
 
     def essays = records.find({
       similarTo(it, "Essays on C.S. Lewis and George MacDonald", "53556908")
@@ -53,6 +58,8 @@ class ModsParsingExamples {
     assert hasIdentifier(essays, "StGlU", "b13803414")
     assert hasIdentifier(essays, "isbn", "0889464944")
 
+    assert essays.creators.size() == 1
+    hasCreator(essays, "Marshall, Cynthia.")
 
     def sketches = records.find({
       similarTo(it, "Statistical sketches of Upper Canada", "69077747")
@@ -62,6 +69,9 @@ class ModsParsingExamples {
 
     assert sketches.identifiers.size() == 1
     assert hasIdentifier(sketches, "UkMaC", "69077747")
+
+    assert sketches.creators.size() == 1
+    hasCreator(sketches, "Dunlop, William")
 
     def mcGuire = records.find({
       similarTo(it, "Edward McGuire, RHA", "22169083")
@@ -74,6 +84,9 @@ class ModsParsingExamples {
     assert hasIdentifier(mcGuire, "UkMaC", "22169083")
     assert hasIdentifier(mcGuire, "StEdNL", "1851914")
 
+    assert mcGuire.creators.size() == 1
+    hasCreator(mcGuire, "Fallon, Brian.")
+
     def influenza = records.find({ similarTo(it,
       "Influenza della Poesia sui Costumi", "43620390") })
 
@@ -81,6 +94,9 @@ class ModsParsingExamples {
 
     assert influenza.identifiers.size() == 1
     assert hasIdentifier(influenza, "UkMaC", "43620390")
+
+    assert influenza.creators.size() == 1
+    hasCreator(influenza, "MABIL, Pier Luigi.")
 
     def nikitovic = records.find({
       similarTo(it, "Pavle Nik Nikitović", "37696876")
@@ -92,6 +108,10 @@ class ModsParsingExamples {
     assert hasIdentifier(nikitovic, "UkMaC", "37696876")
     assert hasIdentifier(nikitovic, "isbn", "8683385124")
 
+    assert nikitovic.creators.size() == 2
+    hasCreator(nikitovic, "Nikitović, Pavle")
+    hasCreator(nikitovic, "Božović, Ratko.")
+
     def grammar = records.find({
       similarTo(it, "Grammaire comparée du grec et du latin", "69250051")
     })
@@ -100,6 +120,10 @@ class ModsParsingExamples {
 
     assert grammar.identifiers.size() == 1
     assert hasIdentifier(grammar, "UkMaC", "69250051")
+
+    assert grammar.creators.size() == 2
+    hasCreator(grammar, "Riemann, Othon.")
+    hasCreator(grammar, "Goelzer, Henri Jules E.")
   }
 
   private void NoTitlesContainEscapedCharacters(records) {
@@ -120,8 +144,12 @@ class ModsParsingExamples {
       record.barcode == expectedBarcode
   }
 
-  private boolean hasIdentifier(record, String namespace, String value) {
-    record.identifiers.any({ it.namespace == namespace && it.value == value })
+  private boolean hasIdentifier(record, String identifierTypeId, String value) {
+    record.identifiers.any({ it.type == identifierTypeId && it.value == value })
+  }
+
+  private void hasCreator(Map<String, Object> instance, String name) {
+    assert instance.creators.any({ it.name == name })
   }
 }