+++ src/test/java/support/fakes/FakeOkapi.java
@@ -32,6 +32,9 @@ public class FakeOkapi extends AbstractVerticle {
     registerFakeMaterialTypesModule(router);
     registerFakeLoanTypesModule(router);
     registerFakeShelfLocationsModule(router);
+    registerFakeInstanceTypesModule(router);
+    registerFakeIdentifierTypesModule(router);
+    registerFakeCreatorTypesModule(router);
 
     server.requestHandler(router::accept)
       .listen(PORT_TO_USE, result -> {
@@ -62,7 +65,15 @@ public class FakeOkapi extends AbstractVerticle {
   }
 
   private void RegisterFakeInstanceStorageModule(Router router) {
-    registerFakeModule(router, "/instance-storage/instances", "instances");
+    ArrayList<String> requiredProperties = new ArrayList<>();
+
+    requiredProperties.add("title");
+    requiredProperties.add("source");
+    requiredProperties.add("instanceTypeId");
+    requiredProperties.add("creators");
+
+    new FakeStorageModule("/instance-storage/instances", "instances",
+      TENANT_ID, requiredProperties).register(router);
   }
 
   private void registerFakeItemsModule(Router router) {
@@ -85,6 +96,18 @@ public class FakeOkapi extends AbstractVerticle {
 
   private void registerFakeShelfLocationsModule(Router router) {
     registerFakeModule(router, "/shelf-locations", "shelflocations");
+  }
+
+  private void registerFakeIdentifierTypesModule(Router router) {
+    registerFakeModule(router, "/identifier-types", "identifierTypes");
+  }
+
+  private void registerFakeInstanceTypesModule(Router router) {
+    registerFakeModule(router, "/instance-types", "instanceTypes");
+  }
+
+  private void registerFakeCreatorTypesModule(Router router) {
+    registerFakeModule(router, "/creator-types", "creatorTypes");
   }
 
   private void registerFakeModule(