+++ NEWS.md
@@ -2,9 +2,13 @@
 
 * Removes `location` property from Item record
 * Introduces `permanentLocation` and `temporaryLocation` properties (which include the name fetched from `shelf-locations`)
-* Provides inventory interface version 3.0
+* Provides inventory interface version 4.0
 * Requires item-storage interface version 4.0
+* Requires instance-storage interface version 3.0
 * Requires shelf-locations interface version 1.0
+* Requires instance-types interface version 1.0
+* Requires identifier-types interface version 1.0
+* Requires creator-types interface version 1.0
 * Adds mod- prefix to names of the built artifacts (FOLIO-813)
 
 ## 5.1.1 2017-09-01